#57 Move credits popup to server selection screen
