+++ android/app/src/main/java/com/example/medcarapp/ServerSelection.java
@@ -1,19 +1,35 @@
 package com.example.medcarapp;
 
+import android.app.Dialog;
+import android.graphics.Color;
+import android.graphics.drawable.ColorDrawable;
 import android.os.Bundle;
+import android.view.View;
 import androidx.appcompat.app.AppCompatActivity;
 import androidx.recyclerview.widget.LinearLayoutManager;
 import androidx.recyclerview.widget.RecyclerView;
 
 public class ServerSelection extends AppCompatActivity {
     RecyclerView rvServerSelection;
+    Dialog creditDialog;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.server_selection);
-
+        creditDialog();
         serverSelection();
+    }
+
+    public void CreditPopup(View v) {
+        creditDialog.setContentView(R.layout.credit_popup);
+        creditDialog.setCanceledOnTouchOutside(true);
+        creditDialog.getWindow().setBackgroundDrawable(new ColorDrawable(Color.TRANSPARENT));
+        creditDialog.show();
+    }
+
+    public void creditDialog(){
+        creditDialog = new Dialog(this);
     }
 
     public void serverSelection(){
