+++ android/app/src/main/res/layout/activity_main.xml
@@ -61,23 +61,4 @@
         app:layout_constraintTop_toBottomOf="@+id/rvAvaliableCars"
         app:layout_constraintVertical_bias="0.0" />
 
-    <com.google.android.material.floatingactionbutton.FloatingActionButton
-        android:id="@+id/btnCredit"
-        android:onClick="CreditPopup"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginStart="16dp"
-        android:layout_marginLeft="16dp"
-        android:layout_marginTop="16dp"
-        android:layout_marginEnd="33dp"
-        android:layout_marginRight="33dp"
-        android:clickable="true"
-        app:backgroundTint="@color/teal_200"
-        app:layout_constraintEnd_toStartOf="@+id/logo"
-        app:layout_constraintHorizontal_bias="1.0"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent"
-        app:srcCompat="@android:drawable/ic_dialog_info" />
-
-
 </androidx.constraintlayout.widget.ConstraintLayout>
